// nextjs/brisk-web-ssr/src/api/ucenter/types.gen.ts
export type CountryCodeResponse = {
    cn?: string;
    code?: string;
    dismiss?: boolean;
    dismissLogin?: boolean;
    en?: string;
    ico?: string;
    mobileCode?: string;
    supportVoice?: boolean;
    weight?: number;
};